Made an attempt at adding materials
New Obstacle class can be applied to areas.

Areas with obstacles are caught by the player. They can inflict speed penalties on enter and exit,
and add drag while the player is inside. You can even layer them on top of each other!

The tradeoff is that I turned the player into an Area2D. For some reason, this wasn't working out
with CharacterBody2D, which inexplicably doesn't have a function to get overlapping areas (I'm used
to 3d Godot where this isn't the case)

I'm not super happy with this. What do you guys think?